Created game timer class
New timer system for our game, treats time as game currency. Player starts with a set amount of time and spends time to perform actions, when time runs out, the game ends. To make a game timer for the player:
PlayerLifeTimer = NewObject<UGameTimer>(this);

To start with 3 days (72 hours) for example,
PlayerLifeTimer -> InitializeTimer(72.0f);

- The rest of the methods to spend or add minutes, hours, get the time, check if time is expired, etc.